refactor: move FileLoadingSettings and FileLoadingSizeLimitResolver to config_resolvers package
