+++ src/quickapp/config_resolvers/file_loading_size_limit_resolver.py
@@ -1,7 +1,7 @@
 from injector import inject
 
-from quickapp.common.file_loading_settings import FileLoadingSettings
 from quickapp.config.application import ApplicationConfig
+from quickapp.config_resolvers.file_loading_settings import FileLoadingSettings
 
 
 @inject
