+++ src/quickapp/dial_core_services/dial_file_service.py
@@ -9,8 +9,8 @@ from aidial_client.types.metadata import FileItem, FileMetadata
 from injector import inject
 from pydantic import BaseModel, ConfigDict
 
-from quickapp.common.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
 from quickapp.common.state_holder import StateHolder
+from quickapp.config_resolvers.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
 
 logger = logging.getLogger(__name__)
